+++ Ra2CsfToolsGUI/Ra2CsfToolsGUI.csproj
@@ -68,7 +68,7 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="Util\CsfFileExtension.cs" />
+    <Compile Include="YamlExtensions\CsfFileExtension.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
